+++ src/main/java/org/tugraz/sysds/runtime/io/WriterTextCSVParallel.java
@@ -75,7 +75,7 @@ public class WriterTextCSVParallel extends WriterTextCSV
 			int blklen = (int)Math.ceil((double)rlen / numThreads);
 			for(int i=0; i<numThreads & i*blklen<rlen; i++) {
 				Path newPath = new Path(path, IOUtilFunctions.getPartFileName(i));
-				tasks.add(new WriteCSVTask(newPath, job, fs, src, i*blklen, (int)Math.min((i+1)*blklen, rlen), csvprops));
+				tasks.add(new WriteCSVTask(newPath, job, fs, src, i*blklen, Math.min((i+1)*blklen, rlen), csvprops));
 			}
 
 			//wait until all tasks have been executed
